Add SaaS backend
This backend works just like the X11 backend, except it starts HamClock
on its own rather than relying on an existing instance. As the name
implies, it is intended for SaaS environments where multiple HamClocks
may be running on a single machine, and running unused instances would
be a significant waste of resources.